refactor: Move repository implementations to persistence/inmemory directory
